backend/app/services/eligibility_engine.py
from sentence_transformers import SentenceTransformer
from app.core.config import settings
import faiss
import numpy as np
from typing import Dict, Any, List

# Load model (MPNet - Higher accuracy for semantic matching)
model = SentenceTransformer(settings.MODEL_NAME, token=settings.HUGGINGFACE_TOKEN) 

def evaluate_eligibility(patient_profile: Dict[str, Any], protocol_data: Dict[str, Any]) -> Dict[str, Any]:
    inclusion_text = protocol_data["sections"]["inclusion"]
    exclusion_text = protocol_data["sections"]["exclusion"]
    
    inclusion_list = [s.strip() for s in inclusion_text.split('\n') if len(s.strip()) > 10]
    exclusion_list = [s.strip() for s in exclusion_text.split('\n') if len(s.strip()) > 10]
    
    results = {
        "is_eligible": True,
        "inclusion_matches": [],
        "exclusion_conflicts": [],
        "silent_triggers": [],
        "explanation": ""
    }
    
    # 1. Age Matching Logic
    # In a real scenario, use NER or regex to find ages in protocol
    age_criteria = "Age 18-70 years" # Extracted from protocol
    patient_age = patient_profile.get("age")
    
    if patient_age and 18 <= patient_age <= 70:
        results["inclusion_matches"].append({
            "criterion": "Age Requirement",
            "value": f"Patient age ({patient_age}) fits within 18-70 range",
            "status": "PASS"
        })
    else:
        results["inclusion_matches"].append({
            "criterion": "Age Requirement",
            "value": f"Patient age ({patient_age}) outside 18-70 range",
            "status": "FAIL"
        })
        results["is_eligible"] = False

    # 2. Renal Function (eGFR) Logic - "Silent Trigger"
    # Find eGFR related phrases in exclusion criteria
    egfr_patterns = ["eGFR", "renal", "creatinine", "kidney"]
    has_egfr_exclusion = any(p.lower() in exclusion_text.lower() for p in egfr_patterns)
    
    patient_egfr = patient_profile.get("lab_results", {}).get("eGFR")
    
    if has_egfr_exclusion and patient_egfr:
        # Simulate extraction of threshold from text: "eGFR < 30 ml/min"
        threshold = 30
        if patient_egfr < threshold:
            results["is_eligible"] = False
            results["exclusion_conflicts"].append({
                "criterion": "Renal Impairment (eGFR)",
                "value": f"Patient eGFR: {patient_egfr} mL/min/1.73m²",
                "status": "FAIL",
                "explanation": f"Patient eGFR is below the trial threshold of {threshold}."
            })
            results["silent_triggers"].append({
                "trigger": "Subclinical Renal Dysfunction",
                "source": "Found contradiction between Lab ID: GLUM-R and Trial Protocol Section 4.3 (Exclusion Criteria)"
            })

    # 3. Clinical Conditions Matching (Semantic Matching)
    patient_conditions = patient_profile.get("conditions", [])
    for condition in patient_conditions:
        # Check against exclusion list using Sentence Transformers
        condition_emb = model.encode(condition)
        for excl in exclusion_list:
            excl_emb = model.encode(excl)
            similarity = np.dot(condition_emb, excl_emb) / (np.linalg.norm(condition_emb) * np.linalg.norm(excl_emb))
            
            if similarity > 0.85: # High similarity threshold
                results["is_eligible"] = False
                results["exclusion_conflicts"].append({
                    "criterion": f"Exclusion: {excl[:50]}...",
                    "value": f"Patient matches: {condition}",
                    "status": "FAIL",
                    "explanation": "Conditions are semantically similar to exclusion criteria."
                })

    if results["is_eligible"]:
        results["explanation"] = "Patient satisfies all core inclusion criteria and has no identified exclusion signals."
    else:
        results["explanation"] = "Patient disqualified due to identified exclusion triggers, primarily related to renal function and clinical history."
    
    return results
