backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
sentence-transformers
faiss-cpu
redis
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
transformers
torch
pydantic-settings
mem0ai
pandas
numpy
openpyxl
